Fixed format string type conversion warnings.

+++ lib/TileCache.h
@@ -32,6 +32,7 @@
 #include <ctime>
 #include <utility> // for make_pair
 #include <iostream>
+#include <inttypes.h>
 
 #ifdef __APPLE__
   #include <sys/time.h> // for gettimeofday
@@ -122,17 +123,16 @@ namespace degate {
   public:
 
     void print_table() const {
-
       printf("Global Image Tile Cache:\n"
-	     "Used memory : %ld bytes\n"
-	     "Max memory  : %ld bytes\n\n"
+	     "Used memory : %llu bytes\n"
+	     "Max memory  : %llu bytes\n\n"
 	     "Holder           | Last access (sec,nsec)    | Amount of memory\n"
 	     "-----------------+---------------------------+------------------------------------\n",
-	     allocated_memory, max_cache_memory);
+	     (long long unsigned)allocated_memory, (long long unsigned)max_cache_memory);
 
       for(cache_t::const_iterator iter = cache.begin(); iter != cache.end(); ++iter) {
 	cache_entry_t const& entry = iter->second;
-	printf("%16p | %12ld.%12ld | %ld M (%ld bytes)\n",
+	printf("%16p | %12ld.%12ld | %u M (%u bytes)\n",
 	       iter->first, entry.first.tv_sec, entry.first.tv_nsec, entry.second/(1024*1024), entry.second);
 	iter->first->print();
       }
